Ensure the initial view still shows even when an EmptyView is provided
The system will always put a ZStack in the view hierarchy, so wrap the initial view in a ZStack to make sure that if an EmptyView is provided as the initial view, we can still hang a task off the subview. EmptyView doesn't render into the view hierarchy in newer versions of SwiftUI, so cannot be used to "hang" tasks from.